[#163] BodsStatementRepository: extend store to await refresh
Documents stored in Elasticsearch are not immediately available for
retrieval, even on the primary shard. This causes an issue when marking
replaced statements, since `Publisher` `publish_new` frequently fails
between `store` and `mark_replaced_statements`.

Extend `BodsStatementRepository` `store` to accept optional parameter to
await the next refresh. This is unfortunate, and it will have to be seen
what the performance impact is when the bulk imports are run.

Note that `wait_for` refresh strategy is used, not `true`, which is
warned about and would be far more likely to cause performance issues.