[#163] EntityQueryBuilder: restrict build_query, etc. by replaced
Restrict `EntityQueryBuilder` `build_query`, `build_fallback_query`.
If the metadata indicates that the document has been replaced, then do
not include it. Otherwise, include it as normal. This also goes for
indexes which haven't yet had the replaced documents marked.

Rubocop example length warnings are disabled, here. The examples are
indeed extremely long, but refactoring them is out-of-scope.

With this patch, documents marked as replaced by
register-transformer-psc get excluded from the search box in
register-v2. The example search `xerox` yields the expecting number of
results, as compared to Elasticsearch. The chain of replaced documents
was also resolved as part of this, and everything seemed as it ought.

However, only 10 results displayed for register-v2, with no next button.
This patch does not attempt to address that, since debugging
`EntityService` `search` showed the correct number of documents both in
`result` and `new_results`, and `current_page`, `records_per_page`, and
`total_count` all seemed to be correct.