[LEI-4] build: install register-sources-oc
Adding this dependency isn't ideal, but it's needed for a migration
script which will use both OC and BODS sources to process LEIs. Adding
it to a transformer instead was considered, but this would require code
to be added to 3 places (PSC, SK, DK), and really, the LEI migration
script is related more closely to BODS rather than specific
transformers. This approach might well be revisited in the future.

https://github.com/openownership/register-sources-oc/issues/14