Revert "[LEI-4] build: example BODS_INDEX"
This was set so that `bin/migrate-lei` script would use the correct
Elasticsearch index. However, in fact the index could be different for
every run, and needs to be set as an environment variable each time. So,
there is no sensible default to use, and anyway, the wildcard syntax is
incorrect.

This reverts commit 906089738c53252404794648dcfd63df0d4c8196.