[register#183] Register::EntityQueryBuilder: fix company number search
There is no longer a `company_number` field. Instead, there is nested
document `identifiers`, which contains an `id` to be searched instead.
To replicate the previous search behaviour, `identifiers` should be
filtered by a `scheme` or `schemeName`. However, it's likely reasonable
to not do so, and instead expand the scope of the search to all
`identifier.id`. Doing so means that searching for LEIs also works, as
well as UTRs, etc. (if they exist).

https://github.com/openownership/register/issues/183