[#189] fix Mappers::ResolverMappings OC ID to be uppercase
Some entities were not reconciled, despite being expected to be
reconciled into a single statement. This appears to be because some
company numbers are inconsistent in their case (i.e UK Scottish `Sc`
vs `SC` prefix), leading to different OpenCorporates IDs being
constructed. However, visiting the URI for a lowercase one redirects to
the uppercase one.

Fix by capitalising all company numbers used to construct OpenCorporates
IDs. I'm not quite sure whether this is enough, especially as no tests
fail despite this change. It's also hard to test this in dev, because of
OpenCorporates IP whitelisting. I'm also unclear as to whether a
migration will be required for existing data in Elasticsearch, or
whether it will get fixed automatically on the new data import running
the transformers.