[#188] note Builders::OwnershipOrControlStatement build fixme
`Builders::OwnershipOrControlStatement` `build` accepts an optional
`replaces_ids` parameter, but it isn't used anywhere within the
function. I attempted to remove it, but that caused multiple tests to
fail, since they appear to be relying on it for some reason with the
numerous stubs… This is unfortunate. I'm disabling Rubocop for this, but
I think it's either a mistake or an undesirable testing approach.