[#189] refactor into IDENTIFIER_NAME_*, IDENTIFIER_SCHEME_* constants
BODS 0.2 and BODS 0.3 specify that an any of id, scheme, schemeName, uri
are required. This is unfortunate, since it's possible for identifiers
to have schemeName but no scheme, leading to lots of matching text
strings within the code.

This doesn't address the specification, but does at least try to start
cleaning up duplication and brittle non-constant usage in the code.