[#171] fix RecordDeserializer deserialize breaks on null JSON values
It's not clear that this was actually incorrect behaviour, since the
absence of a key is not the same as the presence of a key with a null
value. However, importing AM data via static BODS broke because AM data
contains JSON like

    "identifiers": [{"id": null, "scheme": "ARM-TAXID"}]

This was checked against the AM register, and such data was confirmed.

Rather than be unable to import such data, compact all JSON to remove
keys with null values, arrays with null elements, empty strings, empty
objects, and empty arrays. I can't think of any negative effects of
doing such, but it does change the construction of statements. Since
these changes should be trivial (except possibly in the case of empty
strings), and the code for processing static BODS has't yet been run in
production, this seems like a reasonable approach.