[#222] extend Mappers::ResolverMappings, Transformer to add alt names
Mappers::ResolverMappings alternate_names is set to map to an array of
strings. This isn't in keeping with how LEIs were done, where Identifier
structs were used. However, the existing structs for alternate names
don't make room for this additional information, and currently only
names are needed anyway, so rather than risk numerous knock-on effects
with changing the existing structs, I've opted to map to create a list
of strings instead.