[#213] add combine script with ping, list, index, load
This adds a new combine script, to replace Register Files Combiner. It
connects to Redis, lists the BODS files present in the data directory,
indexes those by statement ID, and loads the index into Redis.

Files are iterated in sorted, not filesystem, order, ensuring that
chronological prefix naming is used to encounter statements in the
correct order.

This process is designed to be repeatable, with the script able to be
called multiple times safely. If there are new files, those will be
added to the index, whereas files already seen are not reprocessed.

Redis commands are executed in bulk using pipe mode, which should be
very fast.