[#213] fix Redis protocol error too big inline request
Redis bulk load commands need to be under 64 KB per line (more or less):

https://github.com/redis/redis-rb/issues/1185

https://github.com/redis/redis/blob/07b292af5e86459e435da5986f87bc4f7415d437/src/server.h#L178

If too many IDs are included in the batch file, the lines are too long,
causing an error:

    ERR Protocol error: too big inline request

Fix by chunking IDs written to lines.

What chunk size to use? 1000 IDs works, and leads to lines of
approximately 23 KB. But Register 2 generates statement IDs of around 20
characters in length, whereas BODS 0.2 and BODS 0.3 says they must be
between 32 and 64 chars (the fact that Register 2 doesn't comply with
the minimum is a separate issue). Multiplying by 20/64 gives a target of
around 20 KB, so reducing to 750 IDs should be comfortable, leading to
lines of around 17 KB (as measured with `wc -c`).