[#213] optimise combine index ids, fixing count
Some source files contain duplicate statement IDs. Although these are
made unique in Redis by the use of a set, they needlessly inflate the
size of the array. Fix by deduplicating prior to array construction.

This also caused the count to be off when comparing the cardinality of
the Redis set compared to the logged count, or indeed the logged count
compared to the logged count on the first run.

Note that it's safe to sort uniquely, here, since it's not affecting the
statement order, only unordered statement IDs used for indexing.