[#213] optimise combine to use Grep instead of Jq select in
Previously, Jq was used with slurpfile to set the IDs as a variable, and
then the `select` filter was used with `IN()` to choose only those lines
which were required according to the Redis set difference. This appears
to work fine, although it does not deduplicate within a single file,
only across files.

However, for a 64M test file containing 75541 statements with 50810
unique statements, this took over 8 minutes to run.

Instead, optimise to use Grep instead. The JSON representation is
ensured to be predictable by piping through Jq compact first, which does
not suffer from the `select` `IN()` performance issues. For the same
test file, this took under 4 seconds -- a 120x speed increase.