[#213] optimise Jq to sort keys
This improves deduplication, since statements which are slightly
different in whitespace or key order would also be deduplicated.
Although this is unlikely to be very useful when handling data coming
from a single source, this could be very useful when combining data from
multiple sources. It's theoretically better, so let's do it already.