[#213] optimise combine to deduplicate within files
Previously, Uniq was used to deduplicate consecutive lines. However,
there are many lines in the source files which are duplicated, but
non-consecutively. Where this occurs in a different file, this is
already handled by the Redis deduplication. Where this occured in the
same file, it could not be deduplicated, however, since sorting the file
would change the order of the statements.

Optimise by using AWK to deduplicate each file fully prior to appending
it to the compressed destination file. This requires enough memory, but
these files are relatively small (currently, always less than 64M, and
in new configurations, less than 5M). Doing so enables lines to appear
once and only once within the entirely destination file, still without
requiring the entire destination file to be read into memory.

This approach should also work when appending to existing destination
files -- that is, working with monthly diffs or similar.