[#213] extend combine to support incremental destination files
With some small changes, the existing combine code can support
incremental destination files. This allows the deltas to be stored in a
separate file. Redis maintains a list of all IDs across both the full
and incremental destination files, allowing for deduplication.

It should be possible to join a full file to its incremental files in
order, to obtain a new full file. Since a separate Gzip stream is
written for each source file, combining these files should be as simple
as using Cat.