[#213] optimise combine replacing SUNIONSTORE
Redis SUNIONSTORE doesn't appear to optimise for when the destination
and one of the sources is the same. In fact, there doesn't appear to be
any Redis command which can insert all the members of one set into
another set.

For PSC dataset, a lot of time is spent in SUNIONSTORE, even when there
are no new statements to be appended. This could simply be skipped, but
this cost is paid by everything all the time.

Optimise by writing a custom function to behave like a non-existent
SCOPY, copying all the members of a set into another set. This goes
through a file, so it's not all that quick for small copies, but this is
unlikely to be an issue given how many members each set has.