[#213] add combine-all script
This is useful for combining multiple destination files. It also
repacks them, to make the final compressed file smaller.

Note that this won't be compatible with generating deltas from the
compressed files. For that, and really wherever possible, the original
destination files are preferred. Ability to combine multiple destination
files is needed in order to maintain existing behaviour and offer
single-file downloads in Register, however; hence this script.

Note also that there is no deduplication between multiple destination
files. This is by design, since statement IDs should be globally unique,
not just within a single datasource. (If this isn't the case, we have
many more problems…)