[#251] extend TransformerStream transform to detect company number
Similar to logic in Ingester PSC streaming, attempt to detect the
company number from the payload if it isn't present in the streamed
data.

I'm not sure whether it's more correct to replace the company number
every time, or only if it's not present. I've opted for the latter,
but it's worth noting that Ingester PSC opts for the former. However,
that situation might well be different.