[#32] bump IDENTIFIER_NAME_LEI
This is highly unfortunate, but because of the way statement IDs are
calculated, without this, it's not possible to change LEIs back to how
they were previously. This includes whether temporarily during
migrations, or even to a statement before a spurious LEI got added, or
even to a time when no LEIs were present if they were added after the
creation of the statement.