[#32] rework M20240625 migrations into LEIsAdd, LEIsRemove
This is more flexible, since it allows the LEIs to be removed to be
passed in as a parameter. Not only that, but it allows the new LEIs to
be added first, which means updates won't be rejected for entities which
had no LEIs at some stage of their history (which is pretty much
everything).