optimise Repository store to force refresh
The design of Register pipelines is not optimal for Elasticsearch. When
documents are written, they are required to be visible immediately to
the next section of code, in order to properly mark replaced statements.
Replaced statements tracking using `metadata.replaced` was something
added many months ago, in order to solve various problems with
identifying which version of a statement was the latest. Since BODS 0.2
had no provision for this, this was added separately, without affecting
BODS exports. This also fixed issues with pagination within Register
web. However, it's necessary to be able to have visibility over those
changes within Elasticsearch immediately after they have been written,
which isn't optimal for Elasticsearch, since it's normally an
eventually-consistent database.

An optional refresh parameter was added, in order to allow transformers
to use non-default refresh settings. This is called `await_refresh`
within Register, which in turn uses Elasticsearch `wait_for` parameter.
Which this isn't anywhere as performant as the Elasticsearch default, it
was a compromise, since it waits for the next refresh interval to occur
before returning. It was understood that `Repository` `store` would be
called with batches of records to update, meaning that hopefully it
wouldn't end up having too much of an impact on performance.

In practice, however, it seems that transformers mostly call `store`
within only 1 or 2 records—that is, multiple lines within a file are not
batched. Since the Elasticsearch refresh interval is by default 1s, this
means many operations incur an additional 1-second penalty, whilst
waiting for the update to be visible. However, in fact `store` is called
2 times for each line being processed, which means the minimum
processing time per line is around 2s. In practice, actually many take
even longer.

A further compromise is to change the refresh parameter to `true` within
Elasticsearch. This is not optimal at all, and incurs a much high load
on the Elasticsearch cluster, as well as poorer concurrent operations
and index utilisation. However, doing so greatly increases the speed in
which a transformer can make an update for a line, in some cases being
able to make an update as quickly as 0.2s (so, around 10x) faster.

A benchmark run locally for

  raw_data/source=PSC/year=2024/month=07/day=02/psc-prod-3-2024-07-02-15-53-02-acef0951-5248-4156-b3d1-317c8cdde324

, a file with 389 lines, takes 715s to run with `wait_for` refresh. With
`true` refresh, this takes 175s (with Redis and Elasticsearch indexes
cleared between benchmarks). This results in a speed-up of around 4x.

https://www.elastic.co/guide/en/elasticsearch/reference/current/docs-refresh.html